// src/shared/enum.tsx
export enum CardCategory {
  BROWN,
  GREY,
  BLUE,
  YELLOW,
  RED,
  GREEN,
  PURPLE
}

export enum CountCategory {
  WONDERS = 'wonders',
  TREASURE = 'treasure',
  MILITARY = 'military',
  BLUE = 'civil',
  YELLOW = 'commercial',
  GREEN = 'science',
  PURPLE = 'guilds',
  LEADERS = 'leaders',
  CITIES = 'cities',
  NAVAL = 'naval',
  ISLAND = 'island'
}